+++ src/product.cc
@@ -12,18 +12,17 @@ int Product::next_qualid_ = 1;
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 Product::Ptr Product::Create(Agent* creator, double quantity,
-                             std::string quality, int package_id) {
+                             std::string quality, std::string package_name) {
   if (qualids_.count(quality) == 0) {
     qualids_[quality] = next_qualid_++;
     creator->context()->NewDatum("Products")
         ->AddVal("QualId", qualids_[quality])
         ->AddVal("Quality", quality)
-        ->AddVal("PackageId", package_id)
         ->Record();
   }
 
   // the next lines must come after qual id setting
-  Product::Ptr r(new Product(creator->context(), quantity, quality, package_id));
+  Product::Ptr r(new Product(creator->context(), quantity, quality, package_name));
   r->tracker_.Create(creator);
   return r;
 }
@@ -63,7 +62,7 @@ Product::Ptr Product::Extract(double quantity) {
 
   quantity_ -= quantity;
 
-  Product::Ptr other(new Product(ctx_, quantity, quality_, package_id_));
+  Product::Ptr other(new Product(ctx_, quantity, quality_, package_name_));
   tracker_.Extract(&other->tracker_);
   return other;
 }
@@ -73,36 +72,38 @@ Resource::Ptr Product::ExtractRes(double qty) {
   return boost::static_pointer_cast<Resource>(Extract(qty));
 }
 
-int Product::package_id() {
-  return package_id_;
+std::string Product::package_name() {
+  return package_name_;
 }
 
-void Product::ChangePackageId(int new_package_id) {
-  if (new_package_id == package_id_ || ctx_ == NULL) {
+void Product::ChangePackage(std::string new_package_name) {
+  if (new_package_name == package_name_ || ctx_ == NULL) {
     // no change needed
     return;
   }
-  else if (new_package_id == Package::unpackaged_id()) {
+  else if (new_package_name == Package::unpackaged_name()) {
     // unpackaged has functionally no restrictions
-    package_id_ = new_package_id;
+    package_name_ = new_package_name;
+    tracker_.Package();
     return;
   }
-  Package::Ptr p = ctx_->GetPackageById(new_package_id);
+  Package::Ptr p = ctx_->GetPackage(new_package_name);
   double min = p->fill_min();
   double max = p->fill_max();
   if (quantity_ >= min && quantity_ <= max) {
-    package_id_ = new_package_id;
+    package_name_ = new_package_name;
+    tracker_.Package();
   } else {
     throw ValueError("Product quantity is outside of package fill limits.");
   }
 }
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-Product::Product(Context* ctx, double quantity, std::string quality, int package_id)
+Product::Product(Context* ctx, double quantity, std::string quality, std::string package_name)
     : quality_(quality),
       quantity_(quantity),
       tracker_(ctx, this),
       ctx_(ctx),
-      package_id_(package_id) {}
+      package_name_(package_name) {}
 
 }  // namespace cyclus