+++ src/res_tracker.cc
@@ -64,6 +64,15 @@ void ResTracker::Absorb(ResTracker* absorbed) {
   Record();
 }
 
+void ResTracker::Package() {
+  if (!tracked_) {
+    return;
+  }
+
+  package_name_ = res_->package_name();
+  Record();
+}
+
 void ResTracker::Record() {
   res_->BumpStateId();
   ctx_->NewDatum("Resources")
@@ -74,6 +83,7 @@ void ResTracker::Record() {
       ->AddVal("Quantity", res_->quantity())
       ->AddVal("Units", res_->units())
       ->AddVal("QualId", res_->qual_id())
+      ->AddVal("PackageName", res_->package_name())
       ->AddVal("Parent1", parent1_)
       ->AddVal("Parent2", parent2_)
       ->Record();